app/requirements.txt
Django==3.2
psycopg2-binary>=2.8
graphene_django>=2.0
django-cors-headers>=3.7.0
pyyaml
django-graphql-jwt==0.3.0
backoff==2.1.2
django-graphql-auth
gunicorn
pywebpush
django-imagekit
Pillow
django-compressor
django-compressor-toolkit